fix:allow author to configure form in locations pages
