+++ blocks/form/form.js
@@ -1,5 +1,6 @@
 import { extractFormDefinition as extractSheetDefinition, renderForm as renderDocForm } from './lib/docform.js';
 import decorateUTM from './utm.js';
+import { buildBlock, getMetadata } from '../../scripts/aem.js';
 
 export default async function decorate(block) {
   const { container, formDef } = await extractSheetDefinition(block);
@@ -22,4 +23,16 @@ export default async function decorate(block) {
       });
     }
   }
+}
+
+export function createFormFromMetadata() {
+  const formPath = getMetadata('form-path');
+  const formStyleClass = getMetadata('form-style');
+  const element = { elems: !formPath ? [] : [`<a href="${formPath}"></a>`] };
+  const form = buildBlock(formPath ? 'form' : 'get-a-quote-form', element);
+  if (formPath && formStyleClass) {
+    const classes = formStyleClass.split(',').map((cls) => cls.trim());
+    classes.forEach((cls) => form.classList.add(cls));
+  }
+  return form;
 }
