+++ templates/marketing-location-page/marketing-location-page.js
@@ -7,6 +7,7 @@ import {
 import { div, h3, p } from '../../scripts/dom-helpers.js';
 import { decorateSidebarTemplate } from '../templates.js';
 import { addJsonLd } from '../../scripts/scripts.js';
+import { createFormFromMetadata } from '../../blocks/form/form.js';
 
 function buildMarketingPage(main) {
   const pageSidebar = main.querySelector('div.body-wrapper > div.page-sidebar');
@@ -16,14 +17,7 @@ function buildMarketingPage(main) {
   formSection.classList.add('section');
   formSection.dataset.sectionStatus = 'initialized';
   formSection.style.display = 'none';
-  const formPath = getMetadata('form-path');
-  const formStyleClass = getMetadata('form-style');
-  const element = { elems: !formPath ? [] : [`<a href="${formPath}"></a>`] };
-  const form = buildBlock(formPath ? 'form' : 'get-a-quote-form', element);
-  if (formPath && formStyleClass) {
-    const classes = formStyleClass.split(',').map((cls) => cls.trim());
-    classes.forEach((cls) => form.classList.add(cls));
-  }
+  const form = createFormFromMetadata();
   formSection.prepend(form);
   const locationDiv = div({ class: 'location' });
   const name = getMetadata('name');
